# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nouh LifeOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [todo, in_progress, done, cancelled]
        - in: query
          name: priority
          schema:
            type: string
            enum: [low, medium, high, urgent]
        - in: query
          name: projectId
          schema:
            type: integer
        - in: query
          name: dueDate
          schema:
            type: string
      responses:
        "200":
          description: Tasks list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    put:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [active, completed, on_hold, cancelled]
        - in: query
          name: type
          schema:
            type: string
            enum: [engineering, content, business, personal]
      responses:
        "200":
          description: Projects list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── HABITS ──────────────────────────────────────────────────────
...
components:
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, done, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        projectId:
          type: integer
        dueDate:
          type: string
        completedAt:
          type: string
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, status, priority, tags, createdAt, updatedAt]
    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, done, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        projectId:
          type: integer
        dueDate:
          type: string
        tags:
          type: array
          items:
            type: string
      required: [title]
    UpdateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, done, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        projectId:
          type: integer
        dueDate:
          type: string
        tags:
          type: array
          items:
            type: string
...
    Project:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [engineering, content, business, personal]
        status:
          type: string
          enum: [active, completed, on_hold, cancelled]
        progress:
          type: number
        stack:
          type: string
        githubUrl:
          type: string
        deadline:
          type: string
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, type, status, progress, tags, createdAt, updatedAt]
    CreateProjectBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [engineering, content, business, personal]
        status:
          type: string
          enum: [active, completed, on_hold, cancelled]
        stack:
          type: string
        githubUrl:
          type: string
        deadline:
          type: string
        tags:
          type: array
          items:
            type: string
      required: [title, type]
    UpdateProjectBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [engineering, content, business, personal]
        status:
          type: string
          enum: [active, completed, on_hold, cancelled]
        progress:
          type: number
        stack:
          type: string
        githubUrl:
          type: string
        deadline:
          type: string
        tags:
          type: array
          items:
            type: string

    # ── HABITS ──────────────────────────────────────────────────────
...